chore(2.0): move index to src folder for better management project
